+++ src/app/kitchensink/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { AddCreditButton } from '@/components/tiles/add-credit-button'
-import { TileCreditForm } from '@/components/tiles/tile-credit-form'
+import { TileCreditForm } from '@/components/tiles/add-credit-form'
 import { Area } from '@/components/ui/area'
 import { Section } from '@/components/ui/section'
